pyproject.toml
[project]
name = "online-results"
version = "0.1.0"
description = ""
authors = [
    {name = "kirillbarhatov",email = "k.s.barkhatov@gmail.com"}
]
requires-python = ">=3.14"
dependencies = [
    "google-api-python-client>=2.170.0",
    "google-auth>=2.40.0",
    "python-dotenv>=1.1.0",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
